mcp-server/requirements.txt
# MSIL MCP Server Dependencies
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Resilience
tenacity==8.2.3
circuitbreaker==1.4.0

# Database
asyncpg==0.30.0
sqlalchemy[asyncio]==2.0.36

# Redis
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.11.0

# OpenAI
openai==1.12.0

# Pydantic
pydantic==2.10.5
pydantic-settings==2.7.1

# YAML/JSON
pyyaml==6.0.1
python-json-logger==2.0.7

# OpenAPI
jsonschema==4.17.3

# Utilities
python-dotenv==1.0.1
uuid6==2024.1.12

# Monitoring
prometheus-client==0.19.0
psutil==5.9.8

# CORS
starlette==0.35.1

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
aiosqlite==0.19.0
